commit map & other change
a map was created with agm-maps clearly linked to google maps thanks to the api key.
Implemented the search method for hotels by region and provinces.
Implemented method for creating csv file from an array, converting to json and downloading it.
Modal implementation to display the details of a given structure in a table.